+++ BambooTracker/module/track.cpp
@@ -2,7 +2,9 @@
 #include <utility>
 
 Track::Track(int number, SoundSource source, int channelInSource, int defPattenSize)
-	: attrib_(std::make_unique<TrackAttribute>())
+	: attrib_(std::make_unique<TrackAttribute>()),
+	  effetDisplayWidth_(0)
+
 {	
 	attrib_->number = number;
 	attrib_->source = source;
@@ -23,8 +25,10 @@ Track::Track(const Track& other)
 	attrib_->source = other.attrib_->source;
 	attrib_->channelInSource = other.attrib_->channelInSource;
 
-	patterns_ = other.patterns_;
 	order_ = other.order_;
+	patterns_ = other.patterns_;
+
+	effetDisplayWidth_ = other.effetDisplayWidth_;
 }
 
 TrackAttribute Track::getAttribute() const
@@ -37,7 +41,7 @@ OrderData Track::getOrderData(int order)
 	OrderData res;
 	res.trackAttribute = getAttribute();
 	res.order = order;
-	res.patten = order_.at(order);
+	res.patten = order_.at(static_cast<size_t>(order));
 	return res;
 }
 
@@ -48,19 +52,19 @@ size_t Track::getOrderSize() const
 
 Pattern& Track::getPattern(int num)
 {
-	return patterns_.at(num);
+	return patterns_.at(static_cast<size_t>(num));
 }
 
 Pattern& Track::getPatternFromOrderNumber(int num)
 {
-	return getPattern(order_.at(num));
+	return getPattern(order_.at(static_cast<size_t>(num)));
 }
 
 int Track::searchFirstUneditedUnusedPattern() const
 {
 	for (size_t i = 0; i < patterns_.size(); ++i) {
 		if (!patterns_[i].existCommand() && !patterns_[i].getUsedCount())
-			return i;
+			return static_cast<int>(i);
 	}
 	return -1;
 }
@@ -70,7 +74,7 @@ int Track::clonePattern(int num)
 	int n = searchFirstUneditedUnusedPattern();
 	if (n == -1) return num;
 	else {
-		patterns_.at(n) = patterns_.at(num).clone(n);
+		patterns_.at(static_cast<size_t>(n)) = patterns_.at(static_cast<size_t>(num)).clone(n);
 		return n;
 	}
 }
@@ -79,7 +83,7 @@ std::vector<int> Track::getEditedPatternIndices() const
 {
 	std::vector<int> list;
 	for (size_t i = 0; i < 256; ++i) {
-		if (patterns_[i].existCommand()) list.push_back(i);
+		if (patterns_[i].existCommand()) list.push_back(static_cast<int>(i));
 	}
 	return list;
 }
@@ -97,9 +101,9 @@ std::set<int> Track::getRegisteredInstruments() const
 
 void Track::registerPatternToOrder(int order, int pattern)
 {
-	patterns_.at(pattern).usedCountUp();
-	patterns_.at(order_.at(order)).usedCountDown();
-	order_.at(order) = pattern;
+	patterns_.at(static_cast<size_t>(pattern)).usedCountUp();
+	patterns_.at(static_cast<size_t>(order_.at(static_cast<size_t>(order)))).usedCountDown();
+	order_.at(static_cast<size_t>(order)) = pattern;
 }
 
 void Track::insertOrderBelow(int order)
@@ -109,18 +113,18 @@ void Track::insertOrderBelow(int order)
 
 	if (order == static_cast<int>(order_.size()) - 1) order_.push_back(n);
 	else order_.insert(order_.begin() + order + 1, n);
-	patterns_[n].usedCountUp();
+	patterns_[static_cast<size_t>(n)].usedCountUp();
 }
 
 void Track::deleteOrder(int order)
 {
-	patterns_.at(order_.at(order)).usedCountDown();
+	patterns_.at(static_cast<size_t>(order_.at(static_cast<size_t>(order)))).usedCountDown();
 	order_.erase(order_.begin() + order);
 }
 
 void Track::swapOrder(int a, int b)
 {
-	std::swap(order_.at(a), order_.at((b)));
+	std::swap(order_.at(static_cast<size_t>(a)), order_.at((static_cast<size_t>(b))));
 }
 
 void Track::changeDefaultPatternSize(size_t size)
@@ -128,6 +132,16 @@ void Track::changeDefaultPatternSize(size_t size)
 	for (auto& ptn : patterns_) {
 		ptn.changeSize(size);
 	}
+}
+
+void Track::setEffectDisplayWidth(size_t w)
+{
+	effetDisplayWidth_ = w;
+}
+
+size_t Track::getEffectDisplayWidth() const
+{
+	return effetDisplayWidth_;
 }
 
 void Track::clearUnusedPatterns()